Move the SearchIndexMonitor into ExternalSearch so everything can be in the same file.
