+++ tests/test_external_search.py
@@ -1154,3 +1154,46 @@ class TestSearchIndexCoverageProvider(DatabaseTest):
         eq_(work, record.obj)
         eq_(True, record.transient)
         eq_('There was an error!', record.exception)
+
+
+class TestSearchIndexMonitor(DatabaseTest):
+
+    def test_process_batch(self):
+        index = DummyExternalSearchIndex()
+
+        # Here's a work.
+        work = self._work()
+        work.presentation_ready = True
+
+        # There is no record that it has ever been indexed
+        def _record(work):
+            records = [
+                x for x in work.coverage_records 
+                if x.operation==WorkCoverageRecord.UPDATE_SEARCH_INDEX_OPERATION
+            ]
+            if not records:
+                return None
+            [record] = records
+            return record
+        eq_(None, _record(work))            
+
+        # Here's a Monitor that can index it.
+        monitor = SearchIndexMonitor(self._db, None, "works-index", 
+                                     index_client=index)
+        eq_("Search index update (works)", monitor.service_name)
+
+        # The first time we call process_batch we handle the one and
+        # only work in the database. The ID of that work is returned for
+        # next time.
+        eq_(work.id, monitor.process_batch(0))
+        self._db.commit()
+
+        # The work was added to the search index.
+        eq_([('works', 'work-type', work.id)], index.docs.keys())
+
+        # A WorkCoverageRecord was created for the Work.
+        assert _record(work) is not None
+
+        # The next time we call process_batch, no work is done and the
+        # result is 0, meaning we're done with every work in the system.
+        eq_(0, monitor.process_batch(work.id))
